refactor: accomoodation 필드 추가
-할인 여부, 숙소 타입
