+++ src/main/java/codesquad/team05/domain/accomodation/AccommodationType.java
@@ -1,0 +1,6 @@
+package codesquad.team05.domain.accomodation;
+
+public enum AccommodationType {
+
+    GUESTHOUSE, HOTEL, MOTEL, PENSION, HANOK, APARTMENT
+}
